# unroll dual-based broadcast backwards propagation