@jrevels 2018-02-06
This passes tests. 

Before this PR:

```
| environment | precomputed layers? | SoA enabled? | size | forward time | backward time |
| -----------:|:-------------------:|:------------:|:----:|:------------:|:-------------:|
|        :gpu |        false        |     true     | 512  |  245.61 us   |   858.28 us   |
|        :gpu |        false        |     true     | 1024 |  343.71 us   |   879.01 us   |
|        :gpu |        false        |     true     | 2048 |  736.21 us   |    1.07 ms    |
|        :gpu |        true         |     true     | 512  |  187.16 us   |   430.73 us   |
|        :gpu |        true         |     true     | 1024 |  214.41 us   |   442.07 us   |
|        :gpu |        true         |     true     | 2048 |  427.50 us   |   533.26 us   |
```

after:


```
| environment | precomputed layers? | SoA enabled? | size | forward time | backward time |
| -----------:|:-------------------:|:------------:|:----:|:------------:|:-------------:|
|        :gpu |        false        |     true     | 512  |  241.06 us   |   394.74 us   |
|        :gpu |        false        |     true     | 1024 |  338.07 us   |   412.98 us   |
|        :gpu |        false        |     true     | 2048 |  731.00 us   |    1.01 ms    |
|        :gpu |        true         |     true     | 512  |  181.44 us   |   217.50 us   |
|        :gpu |        true         |     true     | 1024 |  214.13 us   |   235.22 us   |
|        :gpu |        true         |     true     | 2048 |  422.00 us   |   462.80 us   |
```

So it helps a bit for the lower-dimensional cases, but isn't getting us to the performance we expect (backwards pass should not be slower than the forwards pass). 

I'll leave this PR up, anybody feel free to merge if it's sensible to do so (i.e. if other work doesn't fix this problem in a better way).